refactor(components): [menu] use if...else...
